requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1

# v2.0 AI Models
google-generativeai==0.4.1
openai==1.12.0

# v2.0 LangChain & Memory
langchain==0.1.9
langchain-google-genai==0.0.11

# v2.0 Gmail & Google APIs
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# v2.0 SMS (Twilio)
twilio==8.13.0

# v2.0 Agent Optimization
apscheduler==3.10.4

# v2.0 PDF Processing
pdfplumber==0.10.3

# v2.0 Async & WebSocket
aiohttp==3.9.3
websockets==12.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Development
black==24.2.0
flake8==7.0.0
